+++ beatmap/v2/slider.ts
@@ -75,12 +75,12 @@ export class Slider extends Serializable<ISlider> {
             _headTime: this.headTime,
             _headLineIndex: this.headPosX,
             _headLineLayer: this.headPosY,
-            _headCutDirection: this.headCutDirection,
+            _headCutDirection: this.headDirection,
             _headControlPointlengthMultiplier: this.headLengthMultiplier,
             _tailTime: this.tailTime,
             _tailLineIndex: this.tailPosX,
             _tailLineLayer: this.tailPosY,
-            _tailCutDirection: this.tailCutDirection,
+            _tailCutDirection: this.tailDirection,
             _tailControlPointLengthMultiplier: this.tailLengthMultiplier,
             _sliderMidAnchorMode: this.midAnchor,
         };
@@ -150,10 +150,10 @@ export class Slider extends Serializable<ISlider> {
      *
      * **WARNING:** Dot-directional is not recommended, assumes down-directional.
      */
-    get headCutDirection() {
+    get headDirection() {
         return this.data._headCutDirection;
     }
-    set headCutDirection(value: ISlider['_headCutDirection']) {
+    set headDirection(value: ISlider['_headCutDirection']) {
         this.data._headCutDirection = value;
     }
 
@@ -224,10 +224,10 @@ export class Slider extends Serializable<ISlider> {
      *
      * **WARNING:** Dot-directional is not recommended, assumes down-directional.
      */
-    get tailCutDirection() {
+    get tailDirection() {
         return this.data._tailCutDirection;
     }
-    set tailCutDirection(value: ISlider['_tailCutDirection']) {
+    set tailDirection(value: ISlider['_tailCutDirection']) {
         this.data._tailCutDirection = value;
     }
 
@@ -273,7 +273,7 @@ export class Slider extends Serializable<ISlider> {
         return this;
     }
     setDirection(value: ISlider['_headCutDirection']) {
-        this.headCutDirection = value;
+        this.headDirection = value;
         return this;
     }
     setLengthMultiplier(value: ISlider['_headControlPointlengthMultiplier']) {
@@ -293,7 +293,7 @@ export class Slider extends Serializable<ISlider> {
         return this;
     }
     setTailDirection(value: ISlider['_tailCutDirection']) {
-        this.tailCutDirection = value;
+        this.tailDirection = value;
         return this;
     }
     setTailLengthMultiplier(value: ISlider['_tailControlPointLengthMultiplier']) {
@@ -311,44 +311,44 @@ export class Slider extends Serializable<ISlider> {
         if (flipColor) {
             this.colorType = ((1 + this.colorType) % 2) as typeof this.colorType;
         }
-        switch (this.headCutDirection) {
+        switch (this.headDirection) {
             case 2:
-                this.headCutDirection = 3;
+                this.headDirection = 3;
                 break;
             case 3:
-                this.headCutDirection = 2;
+                this.headDirection = 2;
                 break;
             case 6:
-                this.headCutDirection = 7;
+                this.headDirection = 7;
                 break;
             case 7:
-                this.headCutDirection = 6;
+                this.headDirection = 6;
                 break;
             case 4:
-                this.headCutDirection = 5;
+                this.headDirection = 5;
                 break;
             case 5:
-                this.headCutDirection = 4;
+                this.headDirection = 4;
                 break;
         }
-        switch (this.tailCutDirection) {
+        switch (this.tailDirection) {
             case 2:
-                this.tailCutDirection = 3;
+                this.tailDirection = 3;
                 break;
             case 3:
-                this.tailCutDirection = 2;
+                this.tailDirection = 2;
                 break;
             case 6:
-                this.tailCutDirection = 7;
+                this.tailDirection = 7;
                 break;
             case 7:
-                this.tailCutDirection = 6;
+                this.tailDirection = 6;
                 break;
             case 4:
-                this.tailCutDirection = 5;
+                this.tailDirection = 5;
                 break;
             case 5:
-                this.tailCutDirection = 4;
+                this.tailDirection = 4;
                 break;
         }
         if (this.midAnchor) {
@@ -368,8 +368,8 @@ export class Slider extends Serializable<ISlider> {
             this.headPosY < 0 ||
             this.headPosX <= -1000 ||
             this.headPosX >= 1000 ||
-            (this.headCutDirection >= 1000 && this.headCutDirection <= 1360) ||
-            (this.tailCutDirection >= 1000 && this.tailCutDirection <= 1360)
+            (this.headDirection >= 1000 && this.headDirection <= 1360) ||
+            (this.tailDirection >= 1000 && this.tailDirection <= 1360)
         );
     }
 }