+++ converter/customData/objectToV2.ts
@@ -43,20 +43,6 @@ export default function <T extends ICustomDataBase>(
         };
     }
 
-    // events
-    cd._lightID ??= cd.lightID;
-    cd._easing ??= cd.easing;
-    cd._lerpType ??= cd.lerpType;
-    cd._nameFilter ??= cd.nameFilter;
-    cd._rotation ??= cd.rotation;
-    cd._step ??= cd.step;
-    cd._prop ??= cd.prop;
-    cd._speed ??= cd.speed;
-    cd._direction ??= cd.direction;
-    cd._lockPosition ??= cd.lockRotation;
-    cd._direction ??= cd.direction;
-    cd._preciseSpeed ??= cd.speed;
-
     // delete converted customData
     delete cd.color;
     delete cd.coordinates;
@@ -72,18 +58,6 @@ export default function <T extends ICustomDataBase>(
     delete cd.uninteractable;
     delete cd.worldRotation;
     delete cd.animation;
-    delete cd.lightID;
-    delete cd.easing;
-    delete cd.lerpType;
-    delete cd.nameFilter;
-    delete cd.rotation;
-    delete cd.step;
-    delete cd.prop;
-    delete cd.speed;
-    delete cd.direction;
-    delete cd.lockRotation;
-    delete cd.direction;
-    delete cd.speed;
 
     return cd as T;
 }